Refactor to parse inputs
* Extract the logics to another file
* Validate with zod
